requirements.txt
Pillow>=10.0.0
gpiozero>=2.0
spidev>=3.6
spotipy>=2.23.0
vosk>=0.3.45
faster-whisper>=0.10.0
